Updated copyright notice and set-up scripts now work out the location of the set-up directory dynamically.

+++ setup/arch-rpi2/install-samba.sh
@@ -6,7 +6,7 @@
 #    PES provides an interactive GUI for games console emulators
 #    and is designed to work on the Raspberry Pi.
 #
-#    Copyright (C) 2014 Neil Munday (neil@mundayweb.com)
+#    Copyright (C) 2015 Neil Munday (neil@mundayweb.com)
 #
 #    PES is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -22,10 +22,19 @@
 #    along with PES.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-source /home/pi/pes/setup/arch-rpi2/functions.sh
+setupDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+if [ ! -e $setupDir/functions.sh ]; then
+	echo "Error! $setupDir/functions does not exist!"
+	exit 1
+fi
+
+checkFile $setupDir/smb.conf
+
+source $setupDir/functions.sh
 
 run sudo rm -vf /etc/samba/smb.conf.default
-run sudo cp -v /home/pi/pes/setup/arch-rpi2/smb.conf /etc/samba/smb.conf
+run sudo cp -v $setupDir/smb.conf /etc/samba/smb.conf
 run sudo systemctl start smbd.service
 run sudo systemctl start nmbd.service
 run sudo systemctl enable smbd.service
